cptbox: tentatively add seccomp filtering optimization; #380
This change has been tested to work on x64 for a couple of critical
executors. There might exist some loose ends on ARM.

`seccomp` is used to remove the need to return from `ptrace` for all
syscalls that occur, allowing us to selectively filter more dangerous
syscalls (like `open`), while incurring a smaller penalty for frequent,
but harmless, syscalls (like `write`).

There are a number of semantic differences between pure `ptrace` and
`ptrace` + `seccomp`. Notably, where `ptrace` has pre- and post-syscall
events, `ptrace` + `seccomp` only fires a pre-syscall event.
Nonetheless, functionality like switching syscall registers around works
fine in the pre-syscall event.

A smaller distinction as a result of this is that to cancel a syscall, we
need to pass in -1 as the modified syscall, whereas with pure `ptrace` we
would swap the syscall with `getpid` in the post-syscall event. This
distinction is hidden from the Python side (where you can simply pass -1
at all times).

Code that relied both on changing the syscall to perform a different
action *and* changing its return value is now broken when running under
`seccomp`, but I can't think of a reason why anyone would want to do that
-- we certainly dont.